With space access, occupations and bennu-tooklit dependent features.
Ported space classifications edition to angular.
Updated metadata field removal to comply with the new metadata update model, in which information is never lost and added hierarchy in space classifications.
Made Export to Excel adaptable to the available metadata and fixed and issue with export space in which the space itself was not being added to the xls, only its children.
Fixed missing translations and added BUNDLE to FenixEduSpaceConfiguration which points to the String bundle used in spaces.